[SBL-243] DrawingBoardService 의 트랜잭션 및 분산락 범위를 축소시키기 위해 리팩토링
• DrawingBoardService에서 추첨 처리 로직을 DrawingProcessAdapter로 이동
